request::to_request_packets -> packet::to_packets
